# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the PixelAlchemy hazard and relocation decision dashboard
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ZoneColor = typeof ZoneColor[keyof typeof ZoneColor];


export const ZoneColor = {
  Red: 'Red',
  Orange: 'Orange',
  Yellow: 'Yellow',
  Green: 'Green',
} as const;

export type PriorityBucket = typeof PriorityBucket[keyof typeof PriorityBucket];


export const PriorityBucket = {
  Immediate: 'Immediate',
  'Short-term': 'Short-term',
  'Medium-term': 'Medium-term',
  Monitor: 'Monitor',
} as const;

export type DashboardSummaryZoneCounts = {[key: string]: number};

export type DashboardSummaryPriorityCounts = {[key: string]: number};

export interface DashboardSummary {
  region: string;
  total_villages: number;
  total_population: number;
  total_households: number;
  high_risk_villages: number;
  immediate_priority: number;
  population_at_risk: number;
  zone_counts: DashboardSummaryZoneCounts;
  priority_counts: DashboardSummaryPriorityCounts;
}

export interface ZoneMarker {
  village_id: string;
  village_name: string;
  lat: number;
  lon: number;
  zone_color: ZoneColor;
  hazard_score: number;
  population: number;
}

export type VillageExistingHazardZone = typeof VillageExistingHazardZone[keyof typeof VillageExistingHazardZone];


export const VillageExistingHazardZone = {
  High: 'High',
  Moderate: 'Moderate',
  Yellow: 'Yellow',
} as const;

export interface Village {
  village_id: string;
  village_name: string;
  lat: number;
  lon: number;
  population: number;
  households: number;
  existing_hazard_zone: VillageExistingHazardZone;
  historical_incidents: number;
  hazard_score: number;
  zone_color: ZoneColor;
  priority_score: number;
  priority_bucket: PriorityBucket;
}

export type FacilityType = typeof FacilityType[keyof typeof FacilityType];


export const FacilityType = {
  hospital: 'hospital',
  school: 'school',
  water_source: 'water_source',
} as const;

export interface Facility {
  facility_id: string;
  name: string;
  type: FacilityType;
  lat: number;
  lon: number;
}

export type VillageDetail = Village & {
  nearby_facilities: Facility[];
};

export type VillagePriorityScoreBreakdown = {[key: string]: number};

export type VillagePriority = Village & {
  score_breakdown: VillagePriorityScoreBreakdown;
};

export type RelocationSiteLandAvailability = typeof RelocationSiteLandAvailability[keyof typeof RelocationSiteLandAvailability];


export const RelocationSiteLandAvailability = {
  High: 'High',
  Medium: 'Medium',
} as const;

export type RelocationSiteHazardZone = typeof RelocationSiteHazardZone[keyof typeof RelocationSiteHazardZone];


export const RelocationSiteHazardZone = {
  Yellow: 'Yellow',
  Green: 'Green',
} as const;

export type RelocationSiteCapacityStatus = typeof RelocationSiteCapacityStatus[keyof typeof RelocationSiteCapacityStatus];


export const RelocationSiteCapacityStatus = {
  Ready: 'Ready',
  Limited: 'Limited',
  Insufficient: 'Insufficient',
} as const;

export type RelocationSiteScoreBreakdown = {[key: string]: number};

export interface RelocationSite {
  site_id: string;
  site_name: string;
  lat: number;
  lon: number;
  land_availability: RelocationSiteLandAvailability;
  existing_population: number;
  hazard_zone: RelocationSiteHazardZone;
  distance_to_road_km: number;
  distance_to_water_km: number;
  distance_to_healthcare_km: number;
  carrying_capacity: number;
  available_capacity: number;
  capacity_status: RelocationSiteCapacityStatus;
  suitability_score: number;
  rank: number;
  score_breakdown: RelocationSiteScoreBreakdown;
}

export type RelocationResponseCriteriaWeights = {[key: string]: number};

export interface RelocationResponse {
  village: Village;
  criteria_weights: RelocationResponseCriteriaWeights;
  recommendations: RelocationSite[];
}

export type GetVillagesParams = {
search?: string;
zone?: GetVillagesZone;
priority?: GetVillagesPriority;
};

export type GetVillagesZone = typeof GetVillagesZone[keyof typeof GetVillagesZone];


export const GetVillagesZone = {
  Red: 'Red',
  Orange: 'Orange',
  Yellow: 'Yellow',
  Green: 'Green',
} as const;

export type GetVillagesPriority = typeof GetVillagesPriority[keyof typeof GetVillagesPriority];


export const GetVillagesPriority = {
  Immediate: 'Immediate',
  'Short-term': 'Short-term',
  'Medium-term': 'Medium-term',
  Monitor: 'Monitor',
} as const;

export type GetFacilitiesParams = {
type?: GetFacilitiesType;
};

export type GetFacilitiesType = typeof GetFacilitiesType[keyof typeof GetFacilitiesType];


export const GetFacilitiesType = {
  hospital: 'hospital',
  school: 'school',
  water_source: 'water_source',
} as const;