# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the PixelAlchemy hazard and relocation decision dashboard
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryPriorityCounts } from './dashboardSummaryPriorityCounts';
import type { DashboardSummaryZoneCounts } from './dashboardSummaryZoneCounts';

export interface DashboardSummary {
  region: string;
  total_villages: number;
  total_population: number;
  total_households: number;
  high_risk_villages: number;
  immediate_priority: number;
  population_at_risk: number;
  zone_counts: DashboardSummaryZoneCounts;
  priority_counts: DashboardSummaryPriorityCounts;
}