# lib/api-zod/src/generated/types/facility.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the PixelAlchemy hazard and relocation decision dashboard
 * OpenAPI spec version: 0.1.0
 */
import type { FacilityType } from './facilityType';

export interface Facility {
  facility_id: string;
  name: string;
  type: FacilityType;
  lat: number;
  lon: number;
}