# lib/api-zod/src/generated/types/relocationResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the PixelAlchemy hazard and relocation decision dashboard
 * OpenAPI spec version: 0.1.0
 */
import type { RelocationResponseCriteriaWeights } from './relocationResponseCriteriaWeights';
import type { RelocationSite } from './relocationSite';
import type { Village } from './village';

export interface RelocationResponse {
  village: Village;
  criteria_weights: RelocationResponseCriteriaWeights;
  recommendations: RelocationSite[];
}