# lib/api-zod/src/generated/types/relocationSite.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the PixelAlchemy hazard and relocation decision dashboard
 * OpenAPI spec version: 0.1.0
 */
import type { RelocationSiteCapacityStatus } from './relocationSiteCapacityStatus';
import type { RelocationSiteHazardZone } from './relocationSiteHazardZone';
import type { RelocationSiteLandAvailability } from './relocationSiteLandAvailability';
import type { RelocationSiteScoreBreakdown } from './relocationSiteScoreBreakdown';

export interface RelocationSite {
  site_id: string;
  site_name: string;
  lat: number;
  lon: number;
  land_availability: RelocationSiteLandAvailability;
  existing_population: number;
  hazard_zone: RelocationSiteHazardZone;
  distance_to_road_km: number;
  distance_to_water_km: number;
  distance_to_healthcare_km: number;
  carrying_capacity: number;
  available_capacity: number;
  capacity_status: RelocationSiteCapacityStatus;
  suitability_score: number;
  rank: number;
  score_breakdown: RelocationSiteScoreBreakdown;
}