# lib/api-zod/src/generated/types/village.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the PixelAlchemy hazard and relocation decision dashboard
 * OpenAPI spec version: 0.1.0
 */
import type { PriorityBucket } from './priorityBucket';
import type { VillageExistingHazardZone } from './villageExistingHazardZone';
import type { ZoneColor } from './zoneColor';

export interface Village {
  village_id: string;
  village_name: string;
  lat: number;
  lon: number;
  population: number;
  households: number;
  existing_hazard_zone: VillageExistingHazardZone;
  historical_incidents: number;
  hazard_score: number;
  zone_color: ZoneColor;
  priority_score: number;
  priority_bucket: PriorityBucket;
}